v{64,128,256}_intrinsics: fix some implicit conv warnings
w/clang -fsanitize=integer quiets warnings of the form:
implicit conversion from type 'int' of value -80219265 (32-bit, signed)
to type 'uint32_t' (aka 'unsigned int') changed the value to 4214748031
(32-bit, unsigned)
implicit conversion from type 'unsigned char' of value 165 (8-bit,
unsigned) to type 'char' changed the value to -91 (8-bit, signed)

Bug: aomedia:3136
Bug: b/229626362

+++ aom_dsp/simd/v256_intrinsics_x86.h
@@ -545,7 +545,9 @@ SIMD_INLINE v256 v256_max_u8(v256 a, v256 b) { return _mm256_max_epu8(a, b); }
 
 SIMD_INLINE v256 v256_min_s8(v256 a, v256 b) { return _mm256_min_epi8(a, b); }
 
-SIMD_INLINE uint32_t v256_movemask_8(v256 a) { return _mm256_movemask_epi8(a); }
+SIMD_INLINE uint32_t v256_movemask_8(v256 a) {
+  return (uint32_t)_mm256_movemask_epi8(a);
+}
 
 SIMD_INLINE v256 v256_blend_8(v256 a, v256 b, v256 c) {
   return _mm256_blendv_epi8(a, b, c);
